Лента переделана в отдельный сервис. Добавлены классы: FeedController, FeedDaoImpl, FeedDao, FeedService, FeedController. Удалены старые методы для ленты из соответствующих класов.
