+++ src/main/java/ru/yandex/practicum/filmorate/controller/FeedController.java
@@ -1,0 +1,28 @@
+package ru.yandex.practicum.filmorate.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+import ru.yandex.practicum.filmorate.model.Feed;
+import ru.yandex.practicum.filmorate.service.FeedService;
+
+import java.util.Collection;
+
+@Slf4j
+@RestController
+public class FeedController {
+    private final FeedService feedService;
+
+    @Autowired
+    public FeedController(FeedService feedService) {
+        this.feedService = feedService;
+    }
+
+    @GetMapping("/users/{id}/feed")
+    public Collection<Feed> getFeedList(@PathVariable int id) {
+        log.info("Получаем GET запрос к эндпойнту /users/{}/feed", id);
+        return feedService.getFeedList(id);
+    }
+}
