+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.controller;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
@@ -81,12 +80,6 @@ public class UserController {
     public void deleteUserById(@PathVariable int id) {
         log.info("Получен DELETE запрос к эндпоинту /users/{}", id);
         userService.deleteUserById(id);
-    }
-
-    @GetMapping("/{id}/feed")
-    public Collection<Feed> getFeedList(@PathVariable int id) {
-        log.info("Получаем GET запрос к эндпойнту /users/{}/feed", id);
-        return userService.getFeedList(id);
     }
 
     @GetMapping("/{id}/recommendations")
