+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -2,12 +2,11 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.film.dao.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.user.dao.UserStorage;
 import ru.yandex.practicum.filmorate.storage.user.dao.FriendsDao;
+import ru.yandex.practicum.filmorate.storage.user.dao.UserStorage;
 
 import java.util.List;
 
@@ -19,7 +18,7 @@ public class UserService {
     private final FilmStorage filmStorage;
 
     @Autowired
-    public UserService(@Qualifier("userDbStorage") UserStorage userStorage, FriendsDao friendsDao) {
+    public UserService(UserStorage userStorage, FriendsDao friendsDao, FilmStorage filmStorage) {
         this.userStorage = userStorage;
         this.friendsDao = friendsDao;
         this.filmStorage = filmStorage;
@@ -59,10 +58,6 @@ public class UserService {
 
     public void deleteUserById(int id) {
         userStorage.deleteUserById(id);
-    }
-
-    public List<Feed> getFeedList(int id) {
-        return userStorage.getFeedList(id);
     }
 
     public List<Film> getRecommendedFilmsList(int id) {
