+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/dao/impl/LikesDaoImpl.java
@@ -17,11 +17,13 @@ import java.util.List;
 public class LikesDaoImpl implements LikesDao {
     private final JdbcTemplate jdbcTemplate;
     private final UserDbStorage userDbStorage;
+    private final FeedDaoImpl feedDaoImpl;
 
     @Autowired
-    public LikesDaoImpl(JdbcTemplate jdbcTemplate, UserDbStorage userDbStorage) {
+    public LikesDaoImpl(JdbcTemplate jdbcTemplate, UserDbStorage userDbStorage, FeedDaoImpl feedDaoImpl) {
         this.jdbcTemplate = jdbcTemplate;
         this.userDbStorage = userDbStorage;
+        this.feedDaoImpl = feedDaoImpl;
     }
 
     @Override
@@ -29,7 +31,7 @@ public class LikesDaoImpl implements LikesDao {
         User user = userDbStorage.getUserById(userId);
         String sqlQuery = "INSERT INTO films_likes (film_id, user_id) VALUES (?, ?)";
         jdbcTemplate.update(sqlQuery, filmId, userId);
-        userDbStorage.addFeedList(userId, filmId, EventType.LIKE, Operation.ADD);
+        feedDaoImpl.addFeedList(userId, filmId, EventType.LIKE, Operation.ADD);
     }
 
     @Override
@@ -37,7 +39,7 @@ public class LikesDaoImpl implements LikesDao {
         User user = userDbStorage.getUserById(userId);
         String sqlQuery = "DELETE FROM films_likes WHERE user_id = ? AND film_id = ?";
         jdbcTemplate.update(sqlQuery, user.getId(), filmId);
-        userDbStorage.addFeedList(userId, filmId, EventType.LIKE, Operation.REMOVE);
+        feedDaoImpl.addFeedList(userId, filmId, EventType.LIKE, Operation.REMOVE);
     }
 
     @Override
